+++ Assets/Virtual Agents Framework/Samples/Parallel Tasks Sample/Synchronized Tasks/NavMesh.asset
